+++ packages/AI/Recommendations/Engine/src/Engine.ts
@@ -1,4 +1,4 @@
-import { BaseEngine, Metadata, UserInfo, LogStatus, RunView, EntityInfo, LogError, IMetadataProvider } from '@memberjunction/core';
+import { BaseEngine, Metadata, UserInfo, LogStatus, RunView, EntityInfo, LogError, IMetadataProvider, BaseEnginePropertyConfig } from '@memberjunction/core';
 import { ListDetailEntityType, ListEntity, ListEntityType, RecommendationEntity, RecommendationProviderEntity, RecommendationRunEntity } from '@memberjunction/core-entities';
 import { MJGlobal } from '@memberjunction/global';
 import { RecommendationProviderBase } from './ProviderBase';
@@ -19,10 +19,11 @@ export class RecommendationEngineBase extends BaseEngine<RecommendationEngineBas
   }
 
   public async Config(forceRefresh?: boolean, contextUser?: UserInfo, provider?: IMetadataProvider) {
-    const params = [
+    const params: Array<Partial<BaseEnginePropertyConfig>> = [
       {
         PropertyName: '_RecommendationProviders',
         EntityName: 'Recommendation Providers',
+        CacheLocal: true
       },
     ];