+++ packages/Angular/Generic/entity-viewer/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @memberjunction/ng-entity-viewer
 
+## 2.129.0
+
+### Minor Changes
+
+- c7e38aa: migration
+
+### Patch Changes
+
+- Updated dependencies [c391d7d]
+- Updated dependencies [8c412cf]
+- Updated dependencies [fbae243]
+- Updated dependencies [0fb62af]
+- Updated dependencies [7d42aa5]
+- Updated dependencies [c7e38aa]
+- Updated dependencies [7a39231]
+  - @memberjunction/core@2.129.0
+  - @memberjunction/global@2.129.0
+  - @memberjunction/core-entities@2.129.0
+  - @memberjunction/ng-shared-generic@2.129.0
+  - @memberjunction/ng-timeline@2.129.0
+
 ## 2.128.0
 
 ### Patch Changes