+++ packages/MetadataSync/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @memberjunction/metadata-sync
 
+## 2.129.0
+
+### Minor Changes
+
+- 8e2ec79: migration
+
+### Patch Changes
+
+- 20c360f: Fix metadata sync pull operations to correctly process all records and preserve timestamps when data hasn't changed
+- Updated dependencies [c391d7d]
+- Updated dependencies [8c412cf]
+- Updated dependencies [fbae243]
+- Updated dependencies [0fb62af]
+- Updated dependencies [7d42aa5]
+- Updated dependencies [c7e38aa]
+- Updated dependencies [7a39231]
+  - @memberjunction/core@2.129.0
+  - @memberjunction/global@2.129.0
+  - @memberjunction/sqlserver-dataprovider@2.129.0
+  - @memberjunction/graphql-dataprovider@2.129.0
+  - @memberjunction/core-entities@2.129.0
+  - @memberjunction/core-entities-server@2.129.0
+
 ## 2.128.0
 
 ### Patch Changes