+++ packages/QueryGen/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @memberjunction/query-gen
 
+## 2.129.0
+
+### Patch Changes
+
+- Updated dependencies [c391d7d]
+- Updated dependencies [8c412cf]
+- Updated dependencies [fbae243]
+- Updated dependencies [6ce6e67]
+- Updated dependencies [0fb62af]
+- Updated dependencies [7d42aa5]
+- Updated dependencies [c7e38aa]
+- Updated dependencies [7a39231]
+  - @memberjunction/core@2.129.0
+  - @memberjunction/global@2.129.0
+  - @memberjunction/sqlserver-dataprovider@2.129.0
+  - @memberjunction/ai-core-plus@2.129.0
+  - @memberjunction/ai-prompts@2.129.0
+  - @memberjunction/aiengine@2.129.0
+  - @memberjunction/core-entities@2.129.0
+  - @memberjunction/ai-vectors-memory@2.129.0
+  - @memberjunction/ai@2.129.0
+
 ## 2.128.0
 
 ### Patch Changes